+++ apps/charterafrica/src/components/StayInTouch/StayInTouch.js
@@ -3,11 +3,11 @@ import { Link } from "@commons-ui/next";
 import { Grid, SvgIcon } from "@mui/material";
 import React from "react";
 
-import GithubIcon from "@/charterafrica/assets/icons/github.svg";
-import InstagramIcon from "@/charterafrica/assets/icons/instagram.svg";
-import SlackIcon from "@/charterafrica/assets/icons/slack.svg";
 import FacebookIcon from "@/charterafrica/assets/icons/Type=facebook, Size=24, Color=CurrentColor.svg";
+import GithubIcon from "@/charterafrica/assets/icons/Type=github, Size=24, Color=CurrentColor.svg";
+import InstagramIcon from "@/charterafrica/assets/icons/Type=instagram, Size=24, Color=white.svg";
 import LinkedInIcon from "@/charterafrica/assets/icons/Type=linkedin, Size=24, Color=CurrentColor.svg";
+import SlackIcon from "@/charterafrica/assets/icons/Type=slack, Size=24, Color=white.svg";
 import TwitterIcon from "@/charterafrica/assets/icons/Type=twitter, Size=24, Color=CurrentColor.svg";
 
 const mediaToIcon = {
